backend/server/app/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3
